chore(#4): 패키지 구조 변경
