+++ src/main/java/com/modive/adminservice/api/user/dto/res/UserDriveListEventItem.java
@@ -1,0 +1,13 @@
+package com.modive.adminservice.api.user.dto.res;
+
+import lombok.*;
+
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class UserDriveListEventItem {
+    private String type;
+    private Long count;
+}
