+++ src/main/java/com/modive/adminservice/domain/event/dto/EventsByDriveDTO.java
@@ -1,0 +1,15 @@
+package com.modive.adminservice.domain.event.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class EventsByDriveDTO {
+    private String type;
+    private Long count;
+}
