--- src/main/java/com/modive/adminservice/event/dto/EventTotalCntByTypeDTO.java
@@ -1,15 +1,0 @@
-package com.modive.adminservice.event.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class EventTotalCntByTypeDTO {
-    private String reason;
-    private Long count;
-}
