--- src/main/java/com/modive/adminservice/event/dto/EventsByDriveDTO.java
@@ -1,15 +1,0 @@
-package com.modive.adminservice.event.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class EventsByDriveDTO {
-    private String type;
-    private Long count;
-}
