--- src/main/java/com/modive/adminservice/useradmin/dto/res/UserDriveListEventItem.java
@@ -1,13 +1,0 @@
-package com.modive.adminservice.useradmin.dto.res;
-
-import lombok.*;
-
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserDriveListEventItem {
-    private String type;
-    private Long count;
-}
